Fixed some shit, get ready for conflicts!

+++ Two31/Source/Two31/Characters/CultistCharacter.cpp
@@ -383,7 +383,7 @@ bool ACultistCharacter::EquipWeapon(TSubclassOf<AWeapon> Weapon)
 {
 	if (Weapon != NULL)
 	{
-		for (size_t i = 0; i < WeaponSlots.Num(); i++)
+		for (int32 i = 0; i < WeaponSlots.Num(); i++)
 		{
 			if (WeaponSlots[i] == NULL)
 			{
@@ -417,7 +417,7 @@ void ACultistCharacter::SelectWeaponSlot(int index)
 int ACultistCharacter::GetWeaponIndex()
 {
 	int index = -1;
-	for (size_t i = 0; i < WeaponSlots.Num(); i++)
+	for (int32 i = 0; i < WeaponSlots.Num(); i++)
 	{
 		if (CurrentWeapon == WeaponSlots[i])
 			index = i;
@@ -428,7 +428,7 @@ int ACultistCharacter::GetWeaponIndex()
 void ACultistCharacter::Death()
 {
 	AEnemyCharacter::Death();
-	for (size_t i = 0; i < WeaponSlots.Num(); i++)
+	for (int32 i = 0; i < WeaponSlots.Num(); i++)
 	{
 		if (WeaponSlots[i] != NULL)
 			WeaponSlots[i]->Destroy();
